Can be built in 64 bit

+++ src/login_server/bbtable.h
@@ -1,6 +1,6 @@
 
 
-static const unsigned long bbtable[18+1024] =
+static const uint32_t bbtable[18+1024] =
 {
 0x640cded2,
 0xca6cf7cf,
